PHOENIX-6791 Rename differential-test package oracle -> algebra
The `compile.keyspace.oracle` leaf package and `Oracle` class used
"oracle" in the testing-theory sense (a reference implementation for
differential testing), but the name is confusable with Oracle Database.
Rename:

- Package `compile.keyspace.oracle` -> `compile.keyspace.algebra`.
- Class `Oracle` -> `Reference` (entry point: `Reference.extract(...)`).
- Class `OracleTest` -> `ReferenceTest`.

Also move the 5 algebra files from `phoenix-core-client/src/main/java`
to `phoenix-core/src/test/java` (alongside the other 8 algebra files
already there). The code was always test-only — zero production callers
outside the package — and keeping it in the client's main tree bloated
the shipped jar. Consolidating in `phoenix-core/src/test/java` also
keeps the reference model and its consumers (harness, adapter) in one
classpath.

Doc update: `docs/where-optimizer-v2.md` §4 file-tree and §8 section
heading/prose updated to match. One deliberate use of "test oracle" is
kept in §8 prose to explain the differential-testing concept.

Tests:
- ReferenceTest 19/19, DifferentialHarnessTest 5/5,
  HarnessCorpusTest 1/1, WhereOptimizerTest 142/142,
  QueryOptimizerTest 48/48 (1 pre-existing skip).
